Apply colour emphasis, closing the last gap in ppu.h
include/ppu.h has carried a KNOWN GAPS list since the renderer was
written. This empties it.

WHERE THE OLD RATIONALE WAS WRONG. The header argued emphasis "belongs to
the display path, not here", and that is half right. Applying it does
belong there and still happens there. Capturing it cannot: full_palette
rewrites $2001 mid-frame with `tya / and #$E0 / sta $2001`, so emphasis
varies between scanlines of one picture, and a display path reading the
current PPUMASK would paint every row with the last value written. So the
framebuffer widens from uint8_t to uint16_t and carries the three bits
alongside the index - 64 colours x 8 states, which is what the .pal
format's 1536-byte extension stores and what FCEUX and Nintendulator
support, rather than something invented here.

TWO SOURCES DISAGREE, AND THIS PICKS A SIDE, as CLAUDE.md requires.
Mesen's author, in the nesdev thread "Most emulators i've tested don't do
dual emphasis bits right": "it multiplies the corresponding RGB component
by 1.1 and multiplies the other 2 by 0.9. So when all 3 are active, you
get 1.1x0.9x0.9 = 0.891". The COMPOUNDING is taken - each active bit
applies its own factor, so all three set darkens the picture. The 1.1
BOOST is not: NESdev's Colour emphasis page is explicit that boosting the
named channel is RGB PPU behaviour (2C03/04/05, Vs. and PlayChoice) while
the 2C02 only attenuates. Both halves of that decision are commented at
the site with the test that pins them.

The attenuation constant is 0.816328 - lidnariq's terminated-75-ohm
measurement from NESdev's NTSC video page. NOT 0.746, which I had
intended to use until the research contradicted me: that is a constant
inside Bisqwit's NTSC square-wave synthesis, tuned for a composite
decoder, not a hardware measurement. Columns $xE and $xF are exempt -
"$1D black is affected by color emphasis, but $0F black is not".

MEASURED, and one prediction was wrong. Distinct index+emphasis pairs in
full_palette's frame: 456, which is 57 x 8 EXACTLY - every index under
every emphasis state. I predicted 228 on the assumption the ROM pairs
each state with one row group; it redraws the whole grid under all eight.
The comment records the wrong guess rather than quietly asserting the
right number, and the product being exact is itself the assertion: no
combination missing, none spurious.

That measurement also settles a question no source answered. Neither the
wiki nor the research pass could confirm whether emphasis applies to
forced-backdrop output. Every pixel in this ROM comes through that path,
so 456 rather than 57 answers it empirically: it does.

frameDump.masks_the_index_to_six_bits was rewritten, not repaired. It
asserted $41 folds onto $01, and under the new encoding bit 6 IS emphasis,
so $41 legitimately means "index $01, red emphasised". The contract it
tested no longer exists; what it guarded - never indexing past the table -
still matters, so it now uses bits above the whole 9-bit pixel.

Mutation-checked both halves: dropping the capture in the PPU fails five
tests, and applying attenuation once instead of compounding fails the
compounding test alone.

888 -> 896 executed, 0 skipped, 0 failed; CI-visible 371 -> 379.
scan-build clean. The remaining PPU item is the A12 garbage fetches, which
is a bus-accuracy gap rather than a rendering one.

+++ tests/ppu_render_tests.cpp
@@ -210,7 +210,9 @@ std::string test_label()
 // those need entirely different investigations.
 void dump_frames_for_inspection(PPU& ppu, const Config& config, const std::string& label)
 {
-    std::vector<uint8_t> expected(static_cast<size_t>(kWidth) * kHeight);
+    // uint16_t to match the framebuffer's index+emphasis encoding; the
+    // reference is emphasis-free, so the high bits stay zero.
+    std::vector<uint16_t> expected(static_cast<size_t>(kWidth) * kHeight);
     for (int y = 0; y < kHeight; ++y) {
         for (int x = 0; x < kWidth; ++x) {
             expected[static_cast<size_t>(y) * kWidth + x] = reference_pixel(ppu, config, x, y);
@@ -539,7 +541,7 @@ GTEST_TEST(testPPURender, a_nametable_switch_at_dot_0_takes_effect_at_pixel_16)
         ppu.clock();
     }
 
-    const uint8_t* line = &ppu.framebuffer[100 * PPU::screen_width];
+    const uint16_t* line = &ppu.framebuffer[100 * PPU::screen_width];
 
     // Pixels 0-15: the two tiles fetched during line 99's dots 321-336, before
     // the switch. Still screen A.
